CNV-69438: Standardize uploaded dv / pvc file name when mounting an iso or uploading a new one. Fix YAML structure for PVC uploads.
