+++ src/utils/components/DiskModal/utils/helpers.ts
@@ -328,3 +328,17 @@ export const isHotPluggableEnabled = (featureGates: string[]) => {
   }
   return false;
 };
+
+export const convertDataVolumeToTemplate = (
+  dataVolume: V1beta1DataVolume,
+): V1DataVolumeTemplateSpec => ({
+  metadata: dataVolume.metadata,
+  spec: {
+    source: dataVolume.spec.source,
+    sourceRef: dataVolume.spec.sourceRef,
+    storage: {
+      accessModes: dataVolume.spec.storage?.accessModes?.map((mode) => mode as any),
+      resources: dataVolume.spec.storage?.resources,
+    },
+  },
+});
