+++ src/views/virtualmachines/details/tabs/configuration/storage/components/modal/utils.ts
@@ -6,7 +6,6 @@ import DataVolumeModel from '@kubevirt-ui/kubevirt-api/console/models/DataVolume
 import { V1beta1DataVolume } from '@kubevirt-ui/kubevirt-api/containerized-data-importer/models';
 import {
   V1beta1StorageSpec,
-  V1DataVolumeTemplateSpec,
   V1Disk,
   V1VirtualMachine,
   V1VirtualMachineInstance,
@@ -34,7 +33,7 @@ import {
 import { DiskRowDataLayout } from '@kubevirt-utils/resources/vm/utils/disk/constants';
 import { getVMIDevices } from '@kubevirt-utils/resources/vmi';
 import { ARCHITECTURE_LABEL } from '@kubevirt-utils/utils/architecture';
-import { ensurePath, generatePrettyName, isEmpty } from '@kubevirt-utils/utils/utils';
+import { ensurePath, generateUploadDiskName, isEmpty } from '@kubevirt-utils/utils/utils';
 import { getCluster } from '@multicluster/helpers/selectors';
 import { kubevirtK8sGet } from '@multicluster/k8sRequests';
 
@@ -192,7 +191,7 @@ const buildUploadDiskState = (
   uploadFilename: string,
 ): V1DiskFormState => {
   const vmNamespace = getNamespace(vm);
-  const uploadDataVolumeName = `${cdromName}-${generatePrettyName(UPLOAD_SUFFIX)}`;
+  const uploadDataVolumeName = generateUploadDiskName(cdromName, UPLOAD_SUFFIX);
 
   return {
     dataVolumeTemplate: {
@@ -246,17 +245,3 @@ const buildSelectDiskState = (cdromName: string, selectedISO: string): V1DiskFor
     },
   };
 };
-
-export const convertDataVolumeToTemplate = (
-  dataVolume: V1beta1DataVolume,
-): V1DataVolumeTemplateSpec => ({
-  metadata: dataVolume.metadata,
-  spec: {
-    source: dataVolume.spec.source,
-    sourceRef: dataVolume.spec.sourceRef,
-    storage: {
-      accessModes: dataVolume.spec.storage?.accessModes?.map((mode) => mode as any),
-      resources: dataVolume.spec.storage?.resources,
-    },
-  },
-});
